Debug for python3.6
