+++ .github/workflows/examples.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.8]
+        python-version: [3.6]
       fail-fast:  false
       max-parallel: 2
